working in git error

+++ src/main/java/com/grad/akemha/service/SpecializationService.java
@@ -2,12 +2,14 @@ package com.grad.akemha.service;
 
 import com.grad.akemha.dto.specializationDTO.SpecializationRequest;
 import com.grad.akemha.entity.Specialization;
+import com.grad.akemha.exeption.authExceptions.UserNotFoundException;
 import com.grad.akemha.entity.User;
 import com.grad.akemha.repository.SpecializationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class SpecializationService {
@@ -18,9 +20,16 @@ public class SpecializationService {
         return specializationRepository.findAll();
     }
 
-    public void deleteSpecializationById(Long specializationId) {
-        specializationRepository.deleteById(specializationId);
-        return;
+    public Specialization deleteSpecializationById(Long specializationId) {
+        Optional<Specialization> optionalSpecialization = specializationRepository.findById(specializationId);
+        if(optionalSpecialization.isPresent()){
+            Specialization specialization = optionalSpecialization.get();
+            specializationRepository.deleteById(specializationId);
+            return specialization;
+        }else{
+//            throw new NotFoundException("specialization id " + id + " is not found") //TODO
+            throw new UserNotFoundException("specialization id " + specializationId + " is not found");
+        }
     }
 
     public Specialization addSpecialization(SpecializationRequest request) {
